init0
In the name of Allah, the Entirely Merciful, the Especially Merciful.
All praises are for Allah, the Lord of the worlds.
The Most Merciful, the Most Benevolent.
Lord of the day of Judgement.

We only worship You, and we only seek refuge from You.
Guide us the straight path.

The path of those whom You've loved. Never of those who are strayed, and those who are cursed.

Ameen.