The Unlicense.
Enforcing a license never added a single line of code to some of the greatest projects on earth. Enforcing license only did one thing: sue companies and get huge-figured money. Open-source is open anyway! As a Muslim, it does not suit the current maintainer to get huge-figured money that way.